Use a subset of an addon’s package for generating its cache key
`calculateCacheKeyForTree` has been using the entirety of an addon’s `package.json` parsed contents (available as `addonInstance.pkg`) in generating a unique cache key.

The package spec has has reserved properties with `_` and `$` prefixes for package registries to use at their discretion (see http://wiki.commonjs.org/wiki/Packages/1.1). `npm` has begun to take advantage of this by embedding properties that are completely unique to the package (such as its parent). This of course defeats the entire purpose of the cache key.

Therefore, this change moves the implementation to only rely on certain properties that signal the uniqueness of an addon: its version, name, and dependencies (of all forms). This avoids the potential problems with private properties, and should be a bit faster to hash as well.